서비스 계층 수정 (공통 로직 정적 메서드 삭제, 서비스 클래스 통합)
