+++ src/main/java/com/sprint/mission/discodeit/DiscodeitApplication.java
@@ -1,12 +1,10 @@
 package com.sprint.mission.discodeit;
 
 import com.sprint.mission.discodeit.application.dto.request.UserCreateRequestDto;
-import com.sprint.mission.discodeit.application.facade.UserFacade;
 import com.sprint.mission.discodeit.domain.User;
 import com.sprint.mission.discodeit.domain.repository.UserRepository;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.ConfigurableApplicationContext;
 
 import java.io.IOException;
