+++ src/main/java/com/sprint/mission/discodeit/application/BasicUserService.java
@@ -7,15 +7,17 @@ import com.sprint.mission.discodeit.application.dto.request.UserUpdateDto;
 import com.sprint.mission.discodeit.application.dto.response.UserResponseDto;
 import com.sprint.mission.discodeit.domain.BinaryContent;
 import com.sprint.mission.discodeit.domain.User;
-import com.sprint.mission.discodeit.domain.UserStatus;
 import com.sprint.mission.discodeit.domain.repository.UserRepository;
 import com.sprint.mission.discodeit.domain.exception.DuplicateUserException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.UUID;
 
-import static com.sprint.mission.discodeit.application.UserFindHelper.findById;
+
 import static com.sprint.mission.discodeit.application.dto.UserDtoMapper.userToResponseDto;
 
 @Service
@@ -25,7 +27,7 @@ public class BasicUserService{
 
     private final UserRepository userRepository;
 
-    private final FileManager fileManager;
+    private final FileService fileService;
 
 
 
@@ -40,9 +42,9 @@ public class BasicUserService{
                 requestDto.username(),
                 requestDto.phoneNumber());
         userRepository.save(user);
-        fileManager.createUserFolder(user.getId());
+        fileService.createUserFolder(user.getId());
         if (requestDto.profileImage() != null) {
-            BinaryContent content = fileManager.saveUserProfile(user.getId(), requestDto.profileImage());
+            BinaryContent content = fileService.saveUserProfile(user.getId(), requestDto.profileImage());
             user.setProfile(content);
 
         }
@@ -54,7 +56,7 @@ public class BasicUserService{
     }
 
     public UserResponseDto updateUserInfo(UserUpdateDto updateDto) throws IOException {
-        User user = findById(userRepository, updateDto.id());
+        User user = findById(updateDto.id());
         if (updateDto.username() != null) {
             user.updateUsername(updateDto.username());
         }
@@ -65,7 +67,7 @@ public class BasicUserService{
             user.updatePhoneNumber(updateDto.phoneNumber());
         }
         if (updateDto.updateFile() != null) {
-            BinaryContent content = fileManager.saveUserProfile(user.getId(), updateDto.updateFile());
+            BinaryContent content = fileService.saveUserProfile(user.getId(), updateDto.updateFile());
             user.setProfile(content);
         }
         userRepository.save(user);
@@ -74,12 +76,12 @@ public class BasicUserService{
 
 
     public void delete(UserRequestDto requestDto) {
-        userRepository.remove(findById(userRepository,requestDto.id()));
-        fileManager.deleteUserFolder(requestDto.id());
+        userRepository.remove(findById(requestDto.id()));
+        fileService.deleteUserFolder(requestDto.id());
     }
 
     public UserResponseDto getUser(UserRequestDto userRequestDto){
-        User user = findById(userRepository, userRequestDto.id());
+        User user = findById(userRequestDto.id());
         return userToResponseDto(user);
     }
 
@@ -94,5 +96,38 @@ public class BasicUserService{
         userRepository.findByEmail(email).ifPresent(u -> {
             throw new IllegalArgumentException("이미 존재하는 이메일입니다.");
         });
+    }
+
+    public User findById( UUID userId) {
+        return userRepository.findById(userId).orElseThrow(()->new NoSuchElementException("해당 유저가 존재하지 않습니다."));
+    }
+
+    public User findByUsername(String username){
+        return userRepository.findByUsername(username).orElseThrow(()->new NoSuchElementException("아이디가 틀렸습니다"));
+    }
+
+
+    public List<User> findAll(UserRepository userRepository) {
+        return userRepository.findAll();
+    }
+
+    public void login(String loginId, String password) {
+        User user = findByUsername(loginId);
+        if (!user.getPassword().equals(password)){
+            throw new IllegalArgumentException("비밀번호가 틀립니다");
+        }
+        user.markOnline();
+        userRepository.save(user);
+    }
+
+    public void logout(UUID userId){
+        User user = findById(userId);
+        user.markOffline();
+        userRepository.save(user);
+    }
+
+    public boolean checkUserOnline(UUID userId){
+        User user = findById(userId);
+        return user.checkOnline();
     }
 }
