--- src/main/java/com/sprint/mission/discodeit/application/UserStatusService.java
@@ -1,20 +1,0 @@
-package com.sprint.mission.discodeit.application;
-
-import com.sprint.mission.discodeit.domain.User;
-import com.sprint.mission.discodeit.domain.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.util.UUID;
-
-@Service
-@RequiredArgsConstructor
-public class UserStatusService {
-
-    private final UserRepository userRepository;
-
-    public boolean checkUserOnline(UUID userId){
-        User user = UserFindHelper.findById(userRepository,userId);
-        return user.checkOnline();
-    }
-}
